+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package pilz_control
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.5.7 (2019-08-29)
+------------------
+* Add is_executing service to joint trajectory controller
+* Better handling consecutive calls on the hold service of joint trajectory controller
+* Contributors: Pilz GmbH and Co. KG
+
 0.5.6 (2019-06-12)
 ------------------